% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_analysis.R
\name{fisher_test_clusters}
\alias{fisher_test_clusters}
\title{Fisher's exact test}
\usage{
fisher_test_clusters(
  net_clusters,
  int_genes,
  min_size = 10,
  all_genes = NULL,
  filename = NULL
)
}
\arguments{
\item{net_clusters}{Named list of cluster membership for each gene}

\item{int_genes}{List of all genes in category of interest (any not in reference list are ignored)}

\item{min_size}{Min size of cluster to analyse for overlap (default = 10)}

\item{all_genes}{List of all genes to consider as reference list (i.e. full population under analysis, default = NULL in
which case all nodes in net_clusters used)}

\item{filename}{Filename to write results to, default = NULL (no file saved)}
}
\description{
wrapper function for R's fisher.test function to test each cluster identified in graph for overlap with genes of interest
}
