% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{get_cluster_nodes}
\alias{get_cluster_nodes}
\title{Select nodes in cluster list}
\usage{
get_cluster_nodes(net_clusters, select_clusters = NULL, min_size = 10)
}
\arguments{
\item{net_clusters}{Named list of clusters}

\item{select_clusters}{List of cluster IDs to select (default NULL, in which case min_size used to select clusters)}

\item{min_size}{Min size of clusters to include in analysis (default = 10)}
}
\value{
A named logical list, where names are node names, and vector indicates presence of nodes in selection
}
\description{
Takes a network and list of cluster identities (or min size of cluster) and returns a named
logical list indicating whether each gene is in the selected list of clusters
}
