% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{get_combined_network}
\alias{get_combined_network}
\title{Get combined network}
\usage{
get_combined_network(net1, net2, node_list)
}
\arguments{
\item{net1}{Network 1 (igraph format)}

\item{net2}{Network 2 (igraph format)}

\item{node_list}{Vector of names of selected nodes to plot in graph}
}
\description{
Create igraph network with edge weights to indicate shared / unique edges and selected nodes
}
