% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_structure.R
\name{get_communities}
\alias{get_communities}
\title{Detect communities in network}
\usage{
get_communities(
  network,
  filename = NULL,
  cluster_algorithm = "ml",
  min_size = 10
)
}
\arguments{
\item{network}{An igraph representation of network}

\item{filename}{Filename to write summary results about community detection (default = NULL)}

\item{cluster_algorithm}{igraph community detection algorithm ('fg' = fast_greedy,
'ml' = louvain (DEFAULT), 'eb' = edge betweeness, 'wt' = walktrap)}

\item{min_size}{Minimum size of cluster to record details in \code{filename} (default = 20)}
}
\value{
List of community membership for nodes
}
\description{
Uses selected community detection algorithm to find densely connected regions in network
}
