% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_structure.R
\name{get_components}
\alias{get_components}
\title{Check network components}
\usage{
get_components(network, filename = NULL)
}
\arguments{
\item{network}{An igraph representation of network}

\item{filename}{An optional filename (default NULL) to save simple stats about graph and components}
}
\value{
Logical vector indicating which nodes belong to largest component (TRUE) or not (F)
}
\description{
Checks connected components present in network
}
