% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{get_top_nodes}
\alias{get_top_nodes}
\title{Get top nodes}
\usage{
get_top_nodes(degX, cenX, topX)
}
\arguments{
\item{degX}{named vector of node degrees}

\item{cenX}{named vector of node centrality}

\item{topX}{no. of top ranked nodes to select}
}
\value{
index for top nodes in degX/cenX
}
\description{
Get top nodes
}
