% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{go_enrichment}
\alias{go_enrichment}
\title{GO term enrichment of network communities}
\usage{
go_enrichment(
  net_clusters,
  min_size = 10,
  ontology = "geneontology_Biological_Process",
  filename = NULL,
  ref_genes = NULL,
  dirname = NULL
)
}
\arguments{
\item{net_clusters}{Named list of community membership}

\item{min_size}{minimum size of cluster to do GO enrichment for}

\item{ontology}{which ontology to use (any valid WebGestalt enrichDatabase,
default = 'geneontology_Biological_Process')}

\item{filename}{filename to record summarised results}

\item{ref_genes}{list of genes to use as reference for GO enrichment (default = NULL, in which case
full genome used)}

\item{dirname}{directory for WebGestaltR to output results (default = NULL, detailed output suppressed)}
}
\description{
Use WebGestaltR package to do GO term over-representation analysis for each community in network
}
