% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_networks.R
\name{load_pidc_data}
\alias{load_pidc_data}
\title{Loads PIDC data}
\usage{
load_pidc_data(filename)
}
\arguments{
\item{filename}{Path and filename for PIDC .txt output file
(each row is in format: nodeX nodeY edgescore)}
}
\value{
A data.frame with row per edge, columns giving nodeX nodeY score
}
\description{
Loads node and edge data from PIDC output file
}
\details{
Reads in the PIDC output file, removes duplicate edges (network is
undirected so each edge is represented twice in consecutive rows).
}
