% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netbio.R
\docType{package}
\name{netbio}
\alias{netbio}
\alias{netbio-package}
\title{netbio: Analyse biological networks}
\description{
Tools to analyse inferred biological networks
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
