% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_structure.R
\name{plot_community_comparison}
\alias{plot_community_comparison}
\title{Plot to compare network communities}
\usage{
plot_community_comparison(
  net1,
  net2,
  min_size = 10,
  network_names = c("network 1", "network 2")
)
}
\arguments{
\item{net1}{Named list giving community membership for nodes (list names) in network 1}

\item{net2}{Named list giving community membesrhip for nodes in network 2 (must have same nodes as network 1)}

\item{min_size}{Minimum size (no. of nodes) of communities to include in comparison (default = 10)}

\item{network_names}{List giving names for networks to use in plot labels
(default = c("network 1", "network 2"))}
}
\description{
2D heatmap plot to compare (using Jaccard Index) the nodes present in two sets of communities
}
