% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{plot_community_scores}
\alias{plot_community_scores}
\title{Plot to compare network communities}
\usage{
plot_community_scores(network1, network2, net_clusters, min_size = 10)
}
\arguments{
\item{network1}{Data.frame with all network edges (before thresholding),
columns give node1, node2, edge score}

\item{network2}{As above for second network}

\item{net_clusters}{Named list of clusters}

\item{min_size}{Min size of clusters to include in analysis (default = 10)}
}
\description{
Compare edge or node values between networks for communities identified
}
