% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{plot_deg_v_cent}
\alias{plot_deg_v_cent}
\title{Plot single degree v centrality graph}
\usage{
plot_deg_v_cent(degX, cenX, topX, node_list)
}
\arguments{
\item{degX}{named vector of node degrees}

\item{cenX}{named vector of node centrality}

\item{topX}{no. of top ranked nodes to select}

\item{node_list}{Vector of selected node names}
}
\description{
Plot single degree centrality graph for chosen network / node list
}
