% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{plot_degree_centrality}
\alias{plot_degree_centrality}
\title{Plot degree and centrality by node}
\usage{
plot_degree_centrality(net1, net2, node_list, topX = 20)
}
\arguments{
\item{net1}{Network 1 (igraph format)}

\item{net2}{Network 2 (igraph format)}

\item{node_list}{Vector of names of selected nodes to plot in graph}

\item{topX}{No. of top ranked nodes to display names for in graphs}
}
\description{
Various plots of degree and centrality to compare (sub)networks
}
