% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{plot_node_property}
\alias{plot_node_property}
\title{Plot single node property comparison}
\usage{
plot_node_property(p1, p2)
}
\arguments{
\item{p1}{named vector of property in first network}

\item{p2}{named vector of property in second network}
}
\description{
Plot single graph comparing node property (e.g. degree or centrality) values
between two networks
}
