% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{print_shared_properties}
\alias{print_shared_properties}
\title{Print shared node / edge details}
\usage{
print_shared_properties(adj)
}
\arguments{
\item{adj}{igraph adjacency matrix, where edge weights 1, 2, 3 indicate unique edges to networks 1 and 2
and shared edges respectively}
}
\description{
Print message showing number of unique / shared nodes / edges
}
