% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_networks.R
\name{threshold_network}
\alias{threshold_network}
\title{Threshold network}
\usage{
threshold_network(edges, threshold, as_percentage = TRUE, nodes = NULL)
}
\arguments{
\item{edges}{A data.frame with row per edge, columns giving nodeX nodeY score}

\item{threshold}{Threshold, either a \% of top scoring edges to include (default) or a number of edges}

\item{as_percentage}{Logical value, TRUE (default) means \code{threshold} is given as a \%, FALSE indicates
\code{threshold} is the number of edges to keep}

\item{nodes}{Reference list of node names, if not provided (default = NULL), the nodes in the threshold network
are used (in alphabetical order)}
}
\value{
Thresholded network in iGraph format
}
\description{
Creates thresholded network by selecting top scoring edges
}
